voice_assistant/requirements.txt
pyttsx3>=2.90
SpeechRecognition>=3.8.1
wikipedia>=1.4.0
requests>=2.25.1
PyAudio>=0.2.11
python-dotenv>=0.15.0
schedule>=1.1.0
openai>=0.27.0
pvporcupine>=2.1.0
pillow>=8.0.0
psutil>=5.8.0
pyautogui>=0.9.52
pywin32>=300
ctypes-windows-sdk>=0.0.4
# Advanced TTS engines
pygame>=2.1.0
gTTS>=2.2.4
edge-tts>=6.1.1
elevenlabs>=0.2.24
# Background service
pystray>=0.19.0
argparse>=1.4.0
